+++ lib/atlas.rb
@@ -99,7 +99,6 @@ require_relative 'atlas/slot/carrier_efficient'
 
 require_relative 'atlas/runtime'
 require_relative 'atlas/exporter'
-require_relative 'atlas/exporter/full_exporter'
 
 require_relative 'atlas/term/reporter'
